backend-django/banksimple/requirements.txt
Django==1.11.12
djangorestframework==3.8.2
gunicorn==19.7.1
pytz==2018.3
